pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agent-framework>=1.0.0b251120",
    "azure-identity>=1.25.2",
    "python-dotenv>=1.2.1",
]
